api/requirements.txt
certifi==2020.6.20
Django==2.2
djangorestframework==3.9.2
djongo==1.3.3
elasticsearch==7.9.1
pymongo==3.11.0
pytz==2020.1
sqlparse==0.2.4
urllib3==1.25.10
django-rest-swagger==2.1.2
PyYAML==5.3.1